Updating sbt build, renaming and fixing stuff to compile.
