+++ src/main/scala/scala/collection/par/Par.scala
@@ -1,31 +1,6 @@
 package scala.collection.par
 
-
-
-import generic._
-
-
-
 class Par[+Repr](val seq: Repr) {
   override def toString = "Par(%s)".format(seq)
 }
-
-
-trait ParDefs {
-
-  class ops[Repr](val seq: Repr) {
-    def toPar = new Par(seq)
-  }
-
-  implicit def seq2ops[Repr](seq: Repr) = new ops(seq)
-
-  implicit def par2zippable[T, Repr](r: Par[Repr])(implicit isZippable: IsZippable[Repr, T]): Zippable[T] = isZippable(r)
-
-  class assume extends scala.annotation.StaticAnnotation
-
-}
-
-object Par extends ParDefs {
-}
-
