+++ src/main/scala/scala/collection/par/package.scala
